refactor: move error type
